+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Scopes
    - Server Hardware
    - Server Hardware Type
+   - Server Profile Template
    - Storage Pool
    - Storage System
    - Switches